+++ mycodo/mycodo_client.py
@@ -70,6 +70,9 @@ class DaemonControl:
     def get_condition_measurement(self, condition_id):
         return self.rpyc_client.root.get_condition_measurement(condition_id)
 
+    def get_condition_measurement_dict(self, condition_id):
+        return self.rpyc_client.root.get_condition_measurement_dict(condition_id)
+
     def controller_activate(self, controller_type, controller_id):
         return self.rpyc_client.root.controller_activate(
             controller_type, controller_id)